moved motor commands from picarx_improved to motors
